나와 경쟁하기 대기화면 오류 완료, 뱃지 상세보기 하는중...

+++ app/src/main/java/com/team/runnershi/data/AllRecordRecentContent.kt
@@ -3,5 +3,5 @@ package com.team.runnershi.data
 data class AllRecordRecentContent(
     val distance : Int,
     val time : String,
-    val pace : Int
+    val pace : Double
 )
